+++ adapters/Chartboost/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Chartboost iOS Mediation Adapter Changelog
 
+#### Next Version
+- Maps GMA SDK Child flag to Chartboost Consent Status.
+
 #### [Version 9.11.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/chartboost/ChartboostAdapter-9.11.0.0.zip)
 - Verified compatibility with Chartboost SDK version 9.11.0.